Modified home

+++ app/Http/Controllers/PostController.php
@@ -10,6 +10,8 @@ use App\Models\PostUser;
 use Carbon\Carbon;
 
 use App\Notifications\MealReminder;
+use App\Notifications\DeletePost;
+
 
 class PostController extends Controller
 {
@@ -21,7 +23,7 @@ class PostController extends Controller
         Carbon::setLocale('zh-tw');
 
         $user = Auth::user();
-        $posts = Post::all();
+        $posts = Post::orderBy('date', 'asc')->orderBy('time', 'asc')->get();
 
         $userPostIds = PostUser::where('user_id', $user->id)->pluck('post_id')->toArray();
 
@@ -50,10 +52,14 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title' => 'required',
-        ]);
+        $now = Carbon::now();
+        $postDatetime = Carbon::create($request->date . $request->time)->setTimezone('Asia/Taipei');
 
+        if ($postDatetime <= $now) {
+            return redirect()->back()->withErrors(['time' => '不可輸入過去的時間']);
+        }
+
+        //user already create a post at this session
         $existingPost = Post::where('user_id', Auth::user()->id)
             ->where('date', $request->input('date'))
             ->where('time', $request->input('time'))
@@ -63,13 +69,12 @@ class PostController extends Controller
             return redirect()->back()->withErrors(['time' => '您在這個時段已經創建一個飯局了']);
         }
 
-
+        //user already join other post at this session
         $date = $request->input('date');
         $time = $request->input('time');
         $joiningPosts = PostUser::where('user_id', Auth::user()->id)->get();
         foreach ($joiningPosts as $joiningPost) {
             $joinPost = $joiningPost->post;
-            dd($joinPost);
             if ($joinPost->date == $date && $joinPost->time == $time) {
                 return redirect()->back()->withErrors(['time' => '您在這個時段已參加一個飯局了']);
             }
@@ -140,7 +145,15 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->delete();
+        if (auth()->user()->id === $post->user_id) {
+
+            $post_users = PostUser::where('post_id', $post->id)->get();
+            foreach ($post_users as $post_user) {
+                $user = User::find($post_user->user_id);
+                $user->notify(new DeletePost($post));
+            }
+            $post->delete();
+        }
 
         return redirect(route('/'));
     }
@@ -175,5 +188,13 @@ class PostController extends Controller
         }
 
         return redirect()->back();
+    }
+
+    public function endPost(Request $request)
+    {
+        $post = Post::find($request->post_id);
+        $post->status = 2;
+        $post->save();
+        return redirect(route('/'));
     }
 }
